+++ CHANGELOG.md
@@ -1,7 +1,23 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 1.0.11 / 2019-3-26
+- Fix an issue with enviuriarray when paths contained spaces
+
+## 1.0.10 / 2019-2-4
+
+### New Features
+- ENVIURI, ENVIURIARRAY are represented by DEFolder when is_directory is set on the associated parameter
+
+## 1.0.9 / 2018-11-1
+
+### New Features
+- Add datatype support for ENVICoordSys, ENVIRasterSeries, ENVIRasterSeriesArray
+
+## 1.0.8 / 2018-07-10
+
+### New Features
+- Add datatype support for ENVIURIArray, ENVIVirtualizableURIArray, ENVIRasterArray, ENVIROI, ENVIROIArray, SARSCAPEDATA, and SARSCAPEDATAArray
+
 ## 1.0.6 / 2017-11-30
 - Add datatype support for ENVIVirtualizableURI
-
-